# docs/swagger.yaml
basePath: /
definitions:
  model.CommentPaginatedRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
      postID:
        type: integer
    type: object
  model.CreateCommentRequest:
    properties:
      CommentBody:
        type: string
      CommentCreateBy:
        type: string
      postID:
        type: integer
    required:
    - CommentBody
    - CommentCreateBy
    - postID
    type: object
  model.CreateNewsRequest:
    properties:
      postBody:
        type: string
      postHeader:
        type: string
    required:
    - postBody
    - postHeader
    type: object
  model.CreatePostRequest:
    properties:
      postBody:
        type: string
      postCreateBy:
        type: string
      postHeader:
        type: string
    required:
    - postBody
    - postCreateBy
    - postHeader
    type: object
  model.CreateReleaseNoteRequest:
    properties:
      releaseNoteBody:
        type: string
      releaseNoteHeader:
        type: string
    required:
    - releaseNoteBody
    - releaseNoteHeader
    type: object
  model.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  model.UpdateCommentRequest:
    properties:
      commentBody:
        type: string
    required:
    - commentBody
    type: object
  model.UpdateNewsRequest:
    properties:
      postBody:
        type: string
      postHeader:
        type: string
    required:
    - postBody
    - postHeader
    type: object
  model.UpdatePostRequest:
    properties:
      postBody:
        type: string
      postHeader:
        type: string
    required:
    - postBody
    - postHeader
    type: object
host: localhost:8080
info:
  contact: {}
  description: นี่คือ API สำหรับบล็อกพอร์ทัล
  title: API สำหรับบล็อกพอร์ทัล
  version: "1.0"
paths:
  /comments:
    post:
      consumes:
      - application/json
      description: สร้าง comment ใหม่จากข้อมูลที่ส่งมา
      parameters:
      - description: ข้อมูล Comment ที่จะสร้าง
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: สร้าง comment ใหม่
      tags:
      - comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: ลบ comment โดยใช้ ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบ comment ตาม ID
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: อัปเดตข้อมูล comment โดยใช้ ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูล Comment ที่จะอัปเดต
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัปเดต comment ตาม ID
      tags:
      - comments
  /comments/paginated:
    post:
      consumes:
      - application/json
      description: ดึงข้อมูล comment แบบแบ่งหน้าตาม Post ID
      parameters:
      - description: ข้อมูลการแบ่งหน้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CommentPaginatedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูล comment แบบแบ่งหน้า
      tags:
      - comments
  /news:
    post:
      consumes:
      - application/json
      description: สร้างข่าวใหม่ในระบบ
      parameters:
      - description: ข้อมูลข่าวที่จะสร้าง
        in: body
        name: news
        required: true
        schema:
          $ref: '#/definitions/model.CreateNewsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: สร้างข่าวใหม่
      tags:
      - news
  /news/{id}:
    delete:
      consumes:
      - application/json
      description: ลบข่าวตาม ID ที่ระบุ
      parameters:
      - description: ID ของข่าว
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบข่าว
      tags:
      - news
    put:
      consumes:
      - application/json
      description: อัพเดทข้อมูลข่าวตาม ID ที่ระบุ
      parameters:
      - description: ID ของข่าว
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูลข่าวที่จะอัพเดท
        in: body
        name: news
        required: true
        schema:
          $ref: '#/definitions/model.UpdateNewsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัพเดทข่าว
      tags:
      - news
  /posts:
    post:
      consumes:
      - application/json
      description: สร้างโพสต์ใหม่จากข้อมูลที่ส่งมา
      parameters:
      - description: ข้อมูลโพสต์ที่จะสร้าง
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: สร้างโพสต์ใหม่
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: ลบโพสต์ตาม ID ที่ระบุ
      parameters:
      - description: ID ของโพสต์
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบโพสต์
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: อัปเดตข้อมูลโพสต์ตาม ID ที่ระบุ
      parameters:
      - description: ID ของโพสต์
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูลโพสต์ที่จะอัปเดต
        in: body
        name: updatePostRequest
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัปเดตโพสต์
      tags:
      - posts
  /posts/paginated:
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลโพสต์แบบแบ่งหน้าตามพารามิเตอร์ที่ระบุ
      parameters:
      - description: 'หมายเลขหน้า (ค่าเริ่มต้น: 1)'
        in: query
        name: page
        type: integer
      - description: 'จำนวนรายการต่อหน้า (ค่าเริ่มต้น: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลโพสต์แบบแบ่งหน้า
      tags:
      - posts
  /release-notes:
    post:
      consumes:
      - application/json
      description: สร้างบันทึกการเผยแพร่ใหม่ในระบบ
      parameters:
      - description: ข้อมูลบันทึกการเผยแพร่
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateReleaseNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: สร้างบันทึกการเผยแพร่ใหม่
      tags:
      - release-notes
swagger: "2.0"